# feat(video): reduce branching for faster video render task