<!--- Provide a general summary of your changes in the Title above -->

## Description
<!--- Describe your changes in detail -->
Reduce branching in the video task, and replace some ternary operations with bitwise operations and arithmetic instead.

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->
Helps keep the video task performant by reducing the amount of branches that stall the pipeline.

## How has this been tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, tests ran to see how -->
<!--- your change affects other areas of the code, etc. -->
Running `Super Mario Bros.` and a few other roms and letting them auto-play for a little bit to capture the difference repeatably.

### Super Mario Bros.
#### Before
Frames: 1655
FPS: 108.7
Frame Time: [min 2333 us, avg: 9201.9 us, max: 21461 us]
#### After
Frames: 1704
FPS: 111.3
Frame Time: [min 2427 us, avg: 8983.1 us, max: 20611 us]

### Legend of Zelda Links Awakening DX
#### Before
Frames: 1918
FPS: 78.4
Frame Time: [min 2583 us, avg: 12751.7 us, max: 26477 us]
#### After
Frames: 1930
FPS: 79.9
Frame Time: [min 2629 us, avg: 12508.3 us, max: 21521 us]

### Sonic The Hedeghog (Genesis)
#### Before
Frames: 1336
FPS: 49.0
Frame Time: [min 9633 us, avg: 20394.7 us, max: 41146 us]
#### After

Frames: 1336
FPS: 49.4
Frame Time: [min 9545 us, avg: 20255.3 us, max: 42334 us]

## Screenshots (if appropriate, e.g. schematic, board, console logs, lab pictures):

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] Documentation Update
- [ ] Hardware (schematic, board, system design) change
- [x] Software change

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [ ] My change requires a change to the documentation.
- [ ] I have added / updated the documentation related to this change via either README or WIKI

### Software
<!-- Delete this section if not relevant to your PR  -->
- [ ] I have added tests to cover my changes.
- [ ] I have updated the `.github/workflows/build.yml` file to add my new test to the automated cloud build github action.
- [x] All new and existing tests passed.
- [x] My code follows the code style of this project.